the problem was that you were going to the wrong screen
linting

remove extra files
