+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/project-directory/list/filters/filters.scss
