+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/project-directory/show/index.tsx
@@ -1,0 +1,4 @@
+
+export const pathName = '/directory/:projectId';
+
+export default props => "TBD";
